:recycle: chore: Extract FileExplorer and Settings from App
