+++ src/client/file-explorer/hooks/use-file-explorer-component.ts
@@ -2,8 +2,8 @@ import type { OrdoFile, OrdoDirectory } from "@core/app/types"
 
 import { isDirectory } from "@core/app/is-directory"
 
-import Directory from "@client/app/components/file-explorer/directory"
-import File from "@client/app/components/file-explorer/file"
+import Directory from "@client/file-explorer/components/directory"
+import File from "@client/file-explorer/components/file"
 
 export const useFileExplorerComponent = (item: OrdoFile | OrdoDirectory) => {
   return isDirectory(item) ? Directory : File
