+++ src/client/settings/components/checkbox-setting.tsx
