+++ src/client/settings/components/number-setting.tsx
