+++ src/client/settings/components/string-setting.tsx
