+++ src/client/settings/hooks/use-setting-input.ts
@@ -4,11 +4,11 @@ import Switch from "@core/utils/switch"
 
 import { USER_SETTINGS_SCHEMA } from "@core/app/user-settings-schema"
 
-import CheckboxSetting from "@client/app/components/settings/checkbox-setting"
-import SelectSetting from "@client/app/components/settings/select-setting"
-import StringSetting from "@client/app/components/settings/string-setting"
-import NumberSetting from "@client/app/components/settings/number-setting"
-import SelectDirectorySetting from "@client/app/components/settings/select-directory-setting"
+import CheckboxSetting from "@client/settings/components/checkbox-setting"
+import SelectSetting from "@client/settings/components/select-setting"
+import StringSetting from "@client/settings/components/string-setting"
+import NumberSetting from "@client/settings/components/number-setting"
+import SelectDirectorySetting from "@client/settings/components/select-directory-setting"
 import Null from "@client/null"
 import { UserSettings } from "@core/app/types"
 import type { Schema } from "electron-store"
