+++ src/client/settings/index.tsx
@@ -5,7 +5,7 @@ import { USER_SETTINGS_SCHEMA } from "@core/app/user-settings-schema"
 import { disableSideBar } from "@client/app/store"
 import Either from "@core/utils/either"
 
-import SettingsItem from "@client/app/components/settings/settings-item"
+import SettingsItem from "@client/settings/components/settings-item"
 import Null from "@client/null"
 
 /**
