+++ src/client/side-bar/hooks/use-sidebar-component.ts
@@ -2,7 +2,7 @@ import { useAppSelector } from "@client/state"
 import Switch from "@core/utils/switch"
 
 import Null from "@client/null"
-import FileExplorer from "@client/app/file-explorer"
+import FileExplorer from "@client/file-explorer"
 import TagsSidebar from "@client/tags/sidebar"
 
 export const useSidebarComponent = () => {
